Fix serialisation of protocol version identifiers
The exact->inexact strategy only works when the version is a number. The
version is actually a symbol.

Here we introduce explicit strings to use during the serialisation of
the version identifer as the API already  uses a mixture of inexact
numbers and symbols.

This bug was reported by Nick Van Horn when he was trying to use 1.0a
support to bind the Twitter APIs.

Signed-off-by: Andy Bennett <andyjpb@knodium.com>