Fix timestamp and nonce generation
timestamps and nonces are not required for PLAINTEXT signatures but
they have to be valid for HMAC-SHA1 signatures.

Here we do the minimum possible to make them work for the Twitter API.

The nonce should be unique per timestamp but here we just use a
constant. For the Twitter API this seems to work even if more than one
request per second (the granularity of the timestamp) is made.

Signed-off-by: Andy Bennett <andyjpb@knodium.com>